# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Tasks Assignment API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        done:
          type: boolean
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
    UpdateTaskRequest:
      type: object
      required: [done]
      properties:
        done:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
paths:
  /healthz:
    get:
      summary: Healthcheck
      responses:
        '200':
          description: OK
  /v1/tasks:
    get:
      summary: Get tasks or single task by query id
      parameters:
        - in: query
          name: id
          schema:
            type: integer
        - in: query
          name: done
          schema:
            type: boolean
      responses:
        '200':
          description: OK
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
    patch:
      summary: Update task done status
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete task
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
  /v1/external-tasks:
    get:
      summary: Proxy external tasks
      responses:
        '200':
          description: OK